C99 introduced. Loads of warnings fixed.

+++ util/asmotor.h
@@ -55,7 +55,7 @@ extern int log2n(size_t value);
 extern char* _strdup(const char* pszString);
 extern char* _strupr(char* pszString);
 extern char* _strlwr(char* pszString);
-extern int _strnicmp(const char* pszString1, const char* pszString2, int nCount);
+extern int _strnicmp(const char* pszString1, const char* pszString2, size_t nCount);
 extern int _stricmp(const char* pszString1, const char* pszString2);
 #endif
